fixed canvas screen size/zoom issue
Canvas would not load background, start button (with event listeners), or game display properly on smaller screens or when zoomed in/out. Number of pixels/DPI changed in these cases and affected display - kept DPI for scale and changed other usage of canvas width and height to use the offset values.